# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blogidium API Documentation",
        "description": "Comprehensive API documentation for the Blogidium project.",
        "contact": {
            "name": "Nima",
            "email": "nima_8a@yahoo.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/permissions": {
            "get": {
                "tags": [
                    "Roles & Permission"
                ],
                "summary": "List all permissions",
                "operationId": "986fda15b643102ed02bb8f63cde18c5",
                "responses": {
                    "200": {
                        "description": "Permissions retrieved successfully."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles & Permission"
                ],
                "summary": "Create a new permission",
                "operationId": "39f26f188c13388f7c960a23a3b57ff1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the permission",
                                        "type": "string",
                                        "example": "edit-post"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permission created successfully."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/permissions/{permission}": {
            "get": {
                "tags": [
                    "Roles & Permission"
                ],
                "summary": "Get details of a permission",
                "operationId": "afdd7ae1693598988cbbeada23486ab8",
                "parameters": [
                    {
                        "name": "permission",
                        "in": "path",
                        "description": "ID of the permission",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission details retrieved successfully."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Roles & Permission"
                ],
                "summary": "Update a permission",
                "operationId": "9d778dec820137a9e70db6b0284022a3",
                "parameters": [
                    {
                        "name": "permission",
                        "in": "path",
                        "description": "ID of the permission to update",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "New name of the permission",
                                        "type": "string",
                                        "example": "edit-post"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permission updated successfully."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Roles & Permission"
                ],
                "summary": "Delete a permission",
                "operationId": "299adc24adddfaa2ac0ebd9508a314c4",
                "parameters": [
                    {
                        "name": "permission",
                        "in": "path",
                        "description": "ID of the permission to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission deleted successfully."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles & Permission"
                ],
                "summary": "List all roles",
                "operationId": "1823b417ac4516baa67bdc61bf83801b",
                "responses": {
                    "200": {
                        "description": "Roles retrieved successfully."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles & Permission"
                ],
                "summary": "Create a new role",
                "operationId": "879a8fba0a29a9501093d8a3cb51e9e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the role",
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role created successfully."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/roles/{role}": {
            "get": {
                "tags": [
                    "Roles & Permission"
                ],
                "summary": "Get details of a role",
                "operationId": "45d517593492a4f015d67092c3e1dd6f",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "ID of the role",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role details retrieved successfully."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Roles & Permission"
                ],
                "summary": "Update a role",
                "operationId": "c7109408cbde5122b7da9758c8b0fd03",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "ID of the role to update",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "New name of the role",
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role updated successfully."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Roles & Permission"
                ],
                "summary": "Delete a role",
                "operationId": "5aa3a5c9a5c6cf358a5cb8d5d737787e",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "ID of the role to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role deleted successfully."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}